fix var type

+++ sort/core/selection.c
@@ -1,6 +1,6 @@
 #include "sort.h"
 
-void selection_sort(int *arr, int size, int flag)
+void selection_sort(double *arr, int size, int flag)
 {
     int i, j, least, temp;
     for (i = 0; i < size; i++)
